+++ flask_multipass/__init__.py
@@ -13,7 +13,22 @@ from .group import Group
 from .identity import IdentityProvider
 
 
-__version__ = '0.5.5'
 __all__ = ('Multipass', 'AuthProvider', 'IdentityProvider', 'AuthInfo', 'IdentityInfo', 'Group', 'MultipassException',
            'AuthenticationFailed', 'IdentityRetrievalFailed', 'GroupRetrievalFailed', 'NoSuchUser',
            'InvalidCredentials')
+
+
+def __getattr__(name):
+    if name == '__version__':
+        import importlib.metadata
+        import warnings
+
+        warnings.warn(
+            'The `__version__` attribute is deprecated. Use feature detection or'
+            " `importlib.metadata.version('flask-multipass')` instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return importlib.metadata.version('flask-multipass')
+
+    raise AttributeError(name)